Bugs solved

+++ src/main/java/donatehub/domain/projection/DonationInfo.java
@@ -16,5 +16,5 @@ public interface DonationInfo {
 
     String getMessage();
 
-    Integer getAmount();
+    Float getAmount();
 }
